GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch (refs/gitbutler/Virtual-branch)
   branch head: 01eabf3238d0f823b5689ac22f111fef4369a20a
   - vercel.json
   - pages/api/late-notification-scheduler.js
   - pages/api/update.js

Your previous branch was: refs/heads/main

The sha for that commit was: 9f8205f30fadd30782d166af951306121bb758d4

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch